Adding basic MongoDB connectivity
Added basic database connection in config/db.js
that reads in from an ignored config file the password.
I now have a basic server.js file that runs on port 5000
and has a test GET. It seems to be able to connect to
the MongoDB server successfully. I added config as
regular dependency and nodemon as dev dependency.