Adding very basic API keys to the service
Currently, both keys are stored in the uncommitted
config/default.json. The private key can do both POST
and GET, while the public key can only do GET. They
are checked in the new middleware/auth.js and the
appropriate middleware functions are added to the
GET and POST in routes/kitties.js. Note that this isn't
made to be super rigorous, just a basic alpha build.