Eliminating POST from api/kitties path
I deleted the POST path from api/kitties in routes/api/
kitties.js. This is because I think it makes more sense to
have the adding of kitties seperate from the web server,
as it's not necessary for the functionality and it just opens
vulnerabilities. This results in making one security key
redundant, and thus deleted. I changed middleware/
auth.js and routes/api/kitties.js accoundingly. I finally
changed config/db.js to change to a new mongoURI
(now called webAppUri) which signs on with a profile
that can only read from the server. I have a new config/
builder.json file (in .gitignore) that has the URI for signing
in to a profile that has read/write permissions. This will
be used to add kitties to the database, independent of
the web app.