Fixing kitties route
As it is abnormal for GET reqs to have bodies, some
packages will throw an error with the current setup (e.g.
the new dev dependency, node-fetch), I had to change
middleware/auth.js and routes/api/kitties.js to request
the key and emotion from req.query and not req.body.