First attempt at adding HTTPS to kitty-api
I have encapsulated starting the server into
the startServer function in config/startServer.js.
This function allows you to specify a protocol (HTTP
or HTTPS) and read in that setting from the config file.
The config files must now include the 'config' keyword
to be valid. Currently, it does seem to be working fine
locally on Tyr, but I need to test it on the cloud. Finally,
I did update Yarn and Nodemon.