Changing API so it doesn't handle SSL certs
I changed startServer.js so that it simply starts up
a server, listening on the specified PORT and doesn't
look for the http v. https keyword. This should be
handled by the server on which the web app is running.
I also changed around server.js to be consistent with the
changes in startServer.js. I will have to change around
the README file to be consistent with this change.